requirements.txt
asgiref==3.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
black==21.12b0
click==8.0.3
coverage==6.2
Django==4.0.1
django-click==2.3.0
django-cors-headers==3.10.1
django-extensions==3.1.5
djangorestframework==3.13.1
flake8==4.0.1
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
